Fix and clean up type hints
There are cases described in https://github.com/jonase/eastwood#wrong-tag in
which the Clojure compiler will not give an error, but type hints end up
resolving to something unintended.

For instance:

  (defn ^int my-int -2)

will actually result in something more like this:

  (defn ^clojure.core/int my-int -2) ; clojure.core/int is a function!

which is definitely not the intent.

In addition to the fixes described above, all type hints are now immediately
before the argument vector and fully qualified for consistency. Some unnecessary
type hints on locals are removed.