Do not embed constant node ID in compiled code
When a Var is created with the ^:const declaration, the Clojure
compiler will replace all usages of that Var with its value.

This happens at compile time. When loading from source, this has no
impact, but if clj-uuid is ahead-of-time (AOT) compiled then the node
ID is generated once and its value embedded in the compiled artifact.

For example, if an application is AOT-compiled once and deployed on
multiple machines, it will use the same node ID on all machines.

This fixes #37.