replace deprecated primitive-math ns with clj-commons.primitive-math
Single-segment namespaces are not recommended in Clojure
because they translate to the default java package.

This can cause issues with Java interop and with GraalVM compilation
(details https://github.com/clj-easy/graal-build-time/issues/35)

This particular dependency has deprecated the single-segment namespace.
https://github.com/clj-commons/primitive-math/blob/a4ce3cb3effbffd2039880a3ed905af60b5cf153/src/primitive_math.clj#L4

This commit proposes to use the newer, non-single-segment namespace